DepartmentService

+++ src/main/java/com/corse/workshopjavafxjdbc/model/entities/Department.java
@@ -8,23 +8,23 @@ public class Department implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private Integer id;
 
     private String name;
 
     public Department(){
     }
 
-    public Department(Long id, String name){
+    public Department(Integer id, String name){
         this.id = id;
         this.name = name;
     }
 
-    public Long getId(){
+    public Integer getId(){
         return id;
     }
 
-    public void setId(Long id){
+    public void setId(Integer id){
         this.id = id;
     }
